fix: moved types in global types file
