+++ client/src/pages/Register/StepOne.tsx
@@ -3,18 +3,10 @@ import { Control } from "react-hook-form";
 import { Button } from "@/components/ui/button";
 import CustomFormField from "@/components/CustomFormField";
 import { Link } from "react-router-dom";
+import { User } from "@/types/global";
 
 type StepOneProps = {
-  control: Control<
-    {
-      firstName: string;
-      lastName: string;
-      email: string;
-      password: string;
-      company?: string | undefined;
-    },
-    any
-  >;
+  control: Control<User, any>;
 };
 
 export const StepOne = ({ control }: StepOneProps) => {
