+++ client/src/pages/Register/StepTwo.tsx
@@ -1,8 +1,8 @@
 import { motion } from "framer-motion";
 import RadioButtons, { Radio } from "@/components/RadioButtons";
 import { Button } from "@/components/ui/button";
-import { Formula } from "./Register";
 import StepBackButton from "@/components/StepBackButton";
+import { Formula } from "@/types/global";
 
 type StepTwoProps = {
   radios: Radio[];
