+++ client/src/types/global.ts
@@ -1,0 +1,9 @@
+export type User = {
+  firstName: string;
+  lastName: string;
+  email: string;
+  password: string;
+  company?: string | undefined;
+};
+
+export type Formula = "free" | "partner";
